Update Reward ref tests to run both new and old rewards logic
