+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/phase0/rewards/RewardsTestExecutor.java
@@ -24,7 +24,7 @@ import java.util.List;
 import java.util.function.Supplier;
 import tech.pegasys.teku.core.Deltas;
 import tech.pegasys.teku.core.epoch.RewardsAndPenaltiesCalculator;
-import tech.pegasys.teku.core.epoch.RewardsAndPenaltiesStep;
+import tech.pegasys.teku.core.epoch.RewardsAndPenaltiesCalculatorImpl;
 import tech.pegasys.teku.core.epoch.status.ValidatorStatuses;
 import tech.pegasys.teku.datastructures.state.BeaconState;
 import tech.pegasys.teku.ethtests.finder.TestDefinition;
@@ -41,10 +41,33 @@ public class RewardsTestExecutor implements TestExecutor {
 
   @Override
   public void runTest(final TestDefinition testDefinition) throws Throwable {
-    final BeaconState state = loadStateFromSsz(testDefinition, "pre.ssz");
-    final ValidatorStatuses validatorStatuses = ValidatorStatuses.create(state);
+    final BeaconState preState = loadStateFromSsz(testDefinition, "pre.ssz");
+    runStandardTests(testDefinition, preState);
+    runDeprecatedTests(testDefinition, preState);
+  }
+
+  private void runStandardTests(final TestDefinition testDefinition, final BeaconState preState)
+      throws Throwable {
+    final ValidatorStatuses validatorStatuses = ValidatorStatuses.create(preState);
     final RewardsAndPenaltiesCalculator calculator =
-        new RewardsAndPenaltiesCalculator(state, validatorStatuses);
+        testDefinition
+            .getSpec()
+            .getEpochProcessor()
+            .createRewardsAndPenaltiesCalculator(preState, validatorStatuses);
+    runTest(testDefinition, calculator);
+  }
+
+  private void runDeprecatedTests(final TestDefinition testDefinition, final BeaconState preState)
+      throws Throwable {
+    final ValidatorStatuses validatorStatuses = ValidatorStatuses.create(preState);
+    final RewardsAndPenaltiesCalculator calculator =
+        new RewardsAndPenaltiesCalculatorImpl(preState, validatorStatuses);
+    runTest(testDefinition, calculator);
+  }
+
+  private void runTest(
+      final TestDefinition testDefinition, final RewardsAndPenaltiesCalculator calculator)
+      throws Throwable {
     assertDeltas(
         testDefinition,
         "head_deltas.yaml",
@@ -87,7 +110,8 @@ public class RewardsTestExecutor implements TestExecutor {
   }
 
   private Supplier<Deltas> apply(
-      final RewardsAndPenaltiesCalculator calculator, final RewardsAndPenaltiesStep step) {
+      final RewardsAndPenaltiesCalculator calculator,
+      final RewardsAndPenaltiesCalculator.Step step) {
     return () -> calculator.getDeltas(step);
   }
